refactor(server): 모노레포 정리 — client·test 패키지 서버로 통합
- react-native-mcp-client, react-native-mcp-test 패키지 제거
- client·test 소스를 react-native-mcp-server로 이전 (client/, test/)
- package.json·bun.lock·examples e2e yaml 정리
