+++ package.json
@@ -9,10 +9,8 @@
     "document"
   ],
   "scripts": {
-    "build": "bun run build:mcp && bun run build:client && bun run build:test",
+    "build": "bun run build:mcp",
     "build:mcp": "bun run --filter @ohah/react-native-mcp-server build",
-    "build:client": "bun run --filter @ohah/react-native-mcp-client build",
-    "build:test": "bun run --filter @ohah/react-native-mcp-test build",
     "postinstall": "bun run build && node scripts/link-mcp-bin.mjs",
     "dev": "bun scripts/dev.ts",
     "dev:all": "bun run --filter='*' dev",
@@ -28,14 +26,14 @@
     "test:unit": "bun test packages/",
     "test:watch": "bun test --watch",
     "test:coverage": "bun test packages/ --coverage",
-    "test:e2e": "bun run --filter @ohah/react-native-mcp-test build && node packages/react-native-mcp-test/dist/cli.js run examples/demo-app/e2e/all-steps.yaml",
-    "test:e2e:phase2": "bun run --filter @ohah/react-native-mcp-test build && node packages/react-native-mcp-test/dist/cli.js run examples/demo-app/e2e/phase2-flow-control.yaml",
-    "test:e2e:network-mock": "bun run --filter @ohah/react-native-mcp-test build && node packages/react-native-mcp-test/dist/cli.js run examples/demo-app/e2e/network-mock.yaml",
-    "test:e2e:phase3": "bun run --filter @ohah/react-native-mcp-test build && node packages/react-native-mcp-test/dist/cli.js run examples/demo-app/e2e/phase3-clear-state-set-location-copy-paste.yaml",
-    "test:e2e:visual-update": "bun run --filter @ohah/react-native-mcp-test build && node packages/react-native-mcp-test/dist/cli.js run examples/demo-app/e2e/visual-regression-update.yaml",
-    "test:e2e:visual-check": "bun run --filter @ohah/react-native-mcp-test build && node packages/react-native-mcp-test/dist/cli.js run examples/demo-app/e2e/visual-regression-check.yaml",
+    "test:e2e": "bun run build:mcp && node packages/react-native-mcp-server/dist/test/cli.js run examples/demo-app/e2e/all-steps.yaml",
+    "test:e2e:phase2": "bun run build:mcp && node packages/react-native-mcp-server/dist/test/cli.js run examples/demo-app/e2e/phase2-flow-control.yaml",
+    "test:e2e:network-mock": "bun run build:mcp && node packages/react-native-mcp-server/dist/test/cli.js run examples/demo-app/e2e/network-mock.yaml",
+    "test:e2e:phase3": "bun run build:mcp && node packages/react-native-mcp-server/dist/test/cli.js run examples/demo-app/e2e/phase3-clear-state-set-location-copy-paste.yaml",
+    "test:e2e:visual-update": "bun run build:mcp && node packages/react-native-mcp-server/dist/test/cli.js run examples/demo-app/e2e/visual-regression-update.yaml",
+    "test:e2e:visual-check": "bun run build:mcp && node packages/react-native-mcp-server/dist/test/cli.js run examples/demo-app/e2e/visual-regression-check.yaml",
     "setup-mcp": "bun run .cursor/scripts/set-mcp-mise-path.ts",
-    "test:yaml": "bun run --filter @ohah/react-native-mcp-test run",
+    "test:yaml": "bun run build:mcp && node packages/react-native-mcp-server/dist/test/cli.js",
     "docs:dev": "bun run --filter document dev",
     "docs:build": "bun run --filter document build",
     "docs:preview": "bun run --filter document preview",
@@ -44,9 +42,7 @@
   "devDependencies": {
     "@happy-dom/global-registrator": "^20.0.11",
     "@modelcontextprotocol/sdk": "^1.25.0",
-    "@ohah/react-native-mcp-client": "workspace:*",
     "@ohah/react-native-mcp-server": "workspace:*",
-    "@ohah/react-native-mcp-test": "workspace:*",
     "@types/bun": "latest",
     "@types/jest": "^30.0.0",
     "@types/node": "^20.0.0",
