--- packages/react-native-mcp-client/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@ohah/react-native-mcp-client",
-  "version": "0.1.0",
-  "description": "Programmatic client SDK for React Native MCP - typed API for E2E testing",
-  "keywords": [
-    "automation",
-    "e2e",
-    "mcp",
-    "react-native",
-    "testing"
-  ],
-  "license": "MIT",
-  "author": "ohah",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ohah/react-native-mcp.git",
-    "directory": "packages/react-native-mcp-client"
-  },
-  "files": [
-    "dist"
-  ],
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "bunx tsdown",
-    "dev": "bunx tsdown --watch"
-  },
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.25.0",
-    "@ohah/react-native-mcp-server": "workspace:*"
-  },
-  "devDependencies": {
-    "@types/node": "^20.0.0",
-    "tsdown": "0.20.3",
-    "typescript": "^5.9.3"
-  }
-}
