--- packages/react-native-mcp-client/tsdown.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from 'tsdown';
-
-export default defineConfig({
-  entry: 'src/index.ts',
-  format: 'es',
-  outDir: 'dist',
-  platform: 'node',
-  dts: true,
-  fixedExtension: false,
-});
