+++ packages/react-native-mcp-server/package.json
@@ -43,7 +43,15 @@
     "./runtime.js": "./runtime.js",
     "./babel-preset": "./babel-preset.cjs",
     "./babel-plugin-app-registry": "./babel-plugin-app-registry.cjs",
-    "./babel-plugin-inject-testid": "./babel-plugin-inject-testid.cjs"
+    "./babel-plugin-inject-testid": "./babel-plugin-inject-testid.cjs",
+    "./client": {
+      "types": "./dist/client/index.d.ts",
+      "import": "./dist/client/index.js"
+    },
+    "./test": {
+      "types": "./dist/test/index.d.ts",
+      "import": "./dist/test/index.js"
+    }
   },
   "publishConfig": {
     "access": "public"
@@ -62,6 +70,7 @@
     "pixelmatch": "^7.1.0",
     "sharp": "^0.34.5",
     "ws": "^8.18.0",
+    "yaml": "^2.8.0",
     "zod": "^3.25.0"
   },
   "devDependencies": {
@@ -70,7 +79,6 @@
     "@babel/parser": "^7.26.0",
     "@babel/traverse": "^7.26.0",
     "@babel/types": "^7.26.0",
-    "@ohah/react-native-mcp-server": "workspace:*",
     "@types/babel__generator": "^7.6.8",
     "@types/babel__traverse": "^7.28.0",
     "@types/bun": "latest",
