+++ packages/react-native-mcp-server/src/__tests__/client/app-client-clamp.test.ts
@@ -21,7 +21,7 @@ mock.module('@modelcontextprotocol/sdk/client/stdio.js', () => ({
 
 // ─── Import after mocks ─────────────────────────────────────
 
-const { AppClient } = await import('../app-client.js');
+const { AppClient } = await import('../../client/app-client.js');
 
 // ─── Helpers ─────────────────────────────────────────────────
