+++ packages/react-native-mcp-server/src/client/errors.ts
