+++ packages/react-native-mcp-server/src/client/viewport-clamp.ts
