+++ packages/react-native-mcp-server/src/test/reporters/console.ts
