+++ packages/react-native-mcp-server/src/test/reporters/json.ts
