+++ packages/react-native-mcp-server/src/test/reporters/junit.ts
